% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_mb}
\alias{convert_to_mb}
\title{Convert to Mb}
\usage{
convert_to_mb(x, n = 0)
}
\arguments{
\item{x}{Genomic position to convert}

\item{n}{Number of digits to round the value}
}
\value{
A float of the converted value
}
\description{
Convert a genomic position to Mega bp (round to n digits).
This function can be called directly to generate labels in plot scales.
}
\examples{
x_mb <- convert_to_mb(1267356, 2)
x_mb: 1.27

}
