% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_manhattan.R
\name{create_manhattan_background_data}
\alias{create_manhattan_background_data}
\title{Create manhattan background data}
\usage{
create_manhattan_background_data(data, track)
}
\arguments{
\item{data}{Data frame with contig, start, and end for each background rectangle}

\item{track}{Track object for the current plot, generated with the \code{\link{manhattan_track}} function}
}
\value{
A data frame with columns:
contig | start | end | color
}
\description{
Create background data frame for the \code{\link{plot_track_manhattan}} function
}
\examples{
track <- manhattan_track("Fst")
background_data <- create_manhattan_background_data(background_data, track)

}
