% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_manhattan.R
\name{create_manhattan_track_data}
\alias{create_manhattan_track_data}
\title{Create manhattan track data}
\usage{
create_manhattan_track_data(data, track, chromosomes.as.numbers = FALSE)
}
\arguments{
\item{data}{Genomic data (e.g. result of PSASS or RADSex loaded with the \code{\link{load_genome_input}} function)}

\item{track}{Track object for the current plot, generated with the \code{\link{manhattan_track}} function}

\item{chromosomes.as.numbers}{If TRUE, replace chromosome names with numbers for readability (default: FALSE)}
}
\value{
A data frame with columns:
Contig_plot | Position_plot | Metric | Point Color
}
\description{
Create input data frame for the \code{\link{plot_track_manhattan}} function from the genomic data and the track information
}
\examples{
genomic_data <- load_genome_input("psass_window.tsv")
track <- manhattan_track("Fst")

track_data <- create_manhattan_track_data(genomic_data, track)

}
