% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_circos.R
\name{draw_circos}
\alias{draw_circos}
\title{Draw circos plot}
\usage{
draw_circos(
  data,
  contig.lengths,
  tracks,
  highlight = NULL,
  output.file = NULL,
  width = 2400,
  height = 2400,
  res = 120,
  default.color = "grey20",
  default.point.size = 0.25,
  default.ylim = NULL,
  default.bg.color = "white",
  default.bg.highlight.color = "grey80",
  sector.titles.expand = NULL
)
}
\arguments{
\item{data}{Genomic data (e.g. result of PSASS or RADSex loaded with the \code{\link{load_genome_input}} function)}

\item{contig.lengths}{Contig lengths vector (e.g. from the output of the \code{\link{load_genome_input}} function)}

\item{tracks}{List of tracks to plot. Tracks can be generated with the \code{\link{region_track}} function}

\item{highlight}{Vector of names of sectors to highlight}

\item{output.file}{Path to an output file for the generated circos plot, or NULL to plot in the current R device (default: NULL)}

\item{width}{Plot width when plotting to an output file, in pixel (default: 2400)}

\item{height}{Plot height when plotting to an output file, in pixel (default: 2400)}

\item{res}{Image resolution when plotting to an output file, in ppi (default: 120)}

\item{default.color}{Default color for a track when not specified in track data (default: "grey20")}

\item{default.point.size}{Default point size for a track when not specified in track data (default: 0.25)}

\item{default.ylim}{Default y-axis limits for a track when not specified in track data (default: NULL, i.e. infer from data)}

\item{default.bg.color}{Default background color for standard sectors in the track (default: "white")}

\item{default.bg.highlight.color}{Default background color for highlighted sectors in the track (default: "grey80")}

\item{sector.titles.expand}{Manually set the space between sector titles and x-axis as a multiple of ymax (default: NULL)}
}
\description{
Generate a circular plot with multiple tracks for the entire genome
}
\examples{
genomic_data <- load_genome_input("psass_window.tsv")
draw_circos(genomic_data$data, genomic_data$lengths,
            tracks = list(track("Fst", label = expression("F"["ST"])),
                          track(c("Snps_females", "Snps_males"), label = "Pool-specific SNPs", color = c("firebrick2", "dodgerblue3")),
                          track(c("Depth_ratio"), label = "Depth ratio", color = "grey50")),
            output.file = "circos.png")

}
