% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_manhattan_plot.R
\name{draw_manhattan_plot}
\alias{draw_manhattan_plot}
\title{Draw a manhattan plot}
\usage{
draw_manhattan_plot(
  data,
  output.file = NULL,
  width = 14,
  height = 8,
  dpi = 300,
  track = "window_fst",
  lg_numbers = FALSE,
  point.size = 0.5,
  point.palette = c("dodgerblue3", "darkgoldenrod2"),
  background.palette = c("grey85", "grey100"),
  depth.type = "absolute",
  min.depth = 10
)
}
\arguments{
\item{data}{A PSASS data structure obtained with the \code{\link{load_data_files}} function.}

\item{output.file}{Path to an output file in PNG format. If NULL, the plot will be drawn in the default graphic device (default: NULL).}

\item{width}{Width of the output file if specified, in inches (default: 14).}

\item{height}{Height of the output file if specified, in inches (default: 8).}

\item{dpi}{Resolution of the output file if specified, in dpi (default: 300).}

\item{track}{Track to be plotted. Possible values are "position_fst", "window_fst", "window_snp_males", "window_snp_females", "depth_males", "depth_females"
(default: "window_fst").}

\item{point.size}{Size of a point in the plot (default 0.5)}

\item{point.palette}{Color palette for the dots (default c("dodgerblue3", "darkgoldenrod2"))}

\item{background.palette}{Color palette for the background (default c("grey85", "grey100"))}

\item{depth.type}{Type of depth to be plotted, either "absolute" or "relative" (default: "absolute").}

\item{min.depth}{Minimum depth to compute depth ratio.
The ratio for positions with depth lower than this value in either sex will be 1 (default: 10).}
}
\description{
Draw a manhattan plot from the PSASS data
}
\examples{

draw_manhattan_plot(data, track = "window_fst")
}
