% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_circos_window_snp_combined.R
\name{draw_window_snp_combined}
\alias{draw_window_snp_combined}
\title{Circos SNP window track}
\usage{
draw_window_snp_combined(data, ylim = c(0, 1.025 * max(data$Males,
  data$Females) + 0.01), bg.col = "white", point.size = 0.01,
  top.track = FALSE, sector.names = NULL, sector.titles.expand = 1.3,
  sectors = NULL, males.color = "dodgerblue3",
  females.color = "firebrick2")
}
\arguments{
\item{data}{SNP window data frame.}

\item{ylim}{Limits of the y axis (default c(0, 1.025 * max(data$Males, data$Females))).}

\item{bg.col}{Background color for sectors, either a single color or a vector of colors for each sector (default "white").}

\item{point.size}{Size of a point in the plot (default 0.01).}

\item{top.track}{If TRUE, this track will be considered top track of the plot and the x axis will be drawn (default FALSE).}

\item{sector.names}{Vector of contig names obtained with the \code{\link{load_contig_names}} (default NULL).}

\item{males.color}{Color for male-specific SNPs (default "dodgerblue3").}

\item{females.color}{Color for female-specific SNPs (default "firebrick2").}

\item{sex}{Sex to draw the track for, either "Females" or "Males".}

\item{sector.title.expand}{Value controlling the distance between sector titles and the top axis (default 1.3).}
}
\description{
Draws a track on a circos plot for SNP window data. This function is intended for use in the \code{\link{circos_plot}} function.
}
