% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_x_scale}
\alias{generate_x_scale}
\title{Generate x-axis scale}
\usage{
generate_x_scale(region.info, n.ticks = 10)
}
\arguments{
\item{region.info}{A list with three elements: contig name, region start, region end (e.g. output of the \code{\link{parse_region}} function)}

\item{n.ticks}{Number of ticks in the scale}
}
\value{
An object of type scale_x_continuous (from ggplot2) encoding the scale
}
\description{
Generate an x-axis scale with a given number of ticks, automatically converting genomic positions to the best unit
}
\examples{
x_mb <- convert_to_kb(1267356, 2)
x_mb: 1267.36

}
