% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_psass_output.R
\name{load_chromosome_names}
\alias{load_chromosome_names}
\title{Load chromosome names}
\usage{
load_chromosome_names(input_file_path)
}
\arguments{
\item{input_file_path}{Path to the chromosome names file}
}
\value{
A vector with contig IDs as names and chromosome names as values, or NULL if there is no input file
}
\description{
Loads chromosome names from a tabulated file
}
\examples{
chromosomes <- load_chromosome_names("chromosomes_names.tsv")
}
