% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chromosomes_names.R
\name{load_chromosomes_names}
\alias{load_chromosomes_names}
\title{Load a chromosomes names file}
\usage{
load_chromosomes_names(input_file_path)
}
\arguments{
\item{input_file_path}{Path to a chromosomes names file.}
}
\value{
A vector with contigs as names and chromosomes names as values, or NULL if there is no input file
}
\description{
Loads a table of chromosomes names.
}
\examples{
lengths <- load_chromosomes_names("chromosomes_names.tsv")
}
