% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_contig_lengths.R
\name{load_contig_lengths}
\alias{load_contig_lengths}
\title{Load a contig lengths file}
\usage{
load_contig_lengths(input_file_path, chromosomes_names = NULL,
  plot.unplaced = TRUE)
}
\arguments{
\item{input_file_path}{Path to a contig lengths file.}

\item{chromosomes}{A vector of chromosomes names obtained with the \code{\link{load_chromosomes_names}} function (default NULL).}
}
\value{
A list with the following elements:
- lg : lengths of contigs determined to be chromosomes
- unplaced : lengths of contigs determined to be unplaced
- plot : lengths of sectors to be plotted
}
\description{
Loads a table of contig lengths. The function separates the lengths of chromosomes and unplaced contigs
If chromosomes names are provided, they will be used to determine the chromosomes. Otherwise, the function will attempt to detect
chromosomes based on contig names (chromosomes start with LG / Chr / NC).
}
\examples{
lengths <- load_contig_lengths("contig_lengths.tsv", chromosomes_names = chromosomes_names)
}
