% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genome_input.R
\name{load_genome_input}
\alias{load_genome_input}
\title{Load a genome input file}
\usage{
load_genome_input(
  input_file_path,
  chromosomes = NULL,
  detect.chromosomes = TRUE,
  unplaced.label = "Unplaced"
)
}
\arguments{
\item{input_file_path}{Path to a genome input file}

\item{chromosomes}{Vector of chromosome names from \code{\link{load_chromosome_names}} (default: NULL)}

\item{detect.chromosomes}{If TRUE, will consider contigs starting with LG, CH, or NC as chromosomes if no chromosomes were specified (default: TRUE)}

\item{unplaced.label}{Unplaced contigs label (default: "Unplaced")}
}
\value{
A list with two elements: "data" = parsed genome metrics data, "lengths" = lengths of contigs to be plotted
}
\description{
Loads a file containing metrics along a genome. Format: Contig | Position | Length | <Metrics> ...
}
\examples{
window_data <- load_genome_input("psass_window.tsv", chromosomes=chromosomes)
}
