% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scaffold.R
\name{plot_scaffold}
\alias{plot_scaffold}
\title{Plot scaffold}
\usage{
plot_scaffold(scaffold, region = NULL, contig_lengths_file_path = NULL,
  prefix = NULL, window_fst_file_path = NULL,
  position_fst_file_path = NULL, window_snps_file_path = NULL,
  position_snps_file_path = NULL, depth_file_path = NULL,
  chromosomes_names_file_path = NULL, output.file = NULL, width = 12,
  height = 4, dpi = 300, tracks = c("window_fst",
  "window_snp_combined", "depth_ratio"), point.size = 0.5,
  depth.type = "relative", min.depth = 10, depth.ratio.lines = FALSE,
  major.lines.y = TRUE, major.lines.x = FALSE,
  fst.window.color = "grey50", males.color = "dodgerblue3",
  females.color = "firebrick2", alpha.value = 0.25)
}
\arguments{
\item{scaffold}{Name of the scaffold to be plotted. The name can be from the dataset (e.g. "NC_0245486.2") or from the
'chromosomes_names' file (e.g. "LG4").}

\item{region}{A vector specifying the boundaries of the region to be plotted, e.g. c(125000, 250000). If NULL, the entire scaffold
will be plotted (default: NULL).}

\item{contig_lengths_file_path}{Path to a contig lengths file.}

\item{prefix}{Prefix (including full path) to a complete dataset. If prefix is specified, it will be override individual file specifications
such as "window_fst_file_path" (default NULL).}

\item{window_fst_file_path}{Path to a FST window output file (default NULL).}

\item{position_fst_file_path}{Path to a FST positions output file (default NULL).}

\item{window_snps_file_path}{Path to a SNPs window output file (default NULL).}

\item{position_snps_file_path}{Path to a SNPs positions output file (default NULL).}

\item{depth_file_path}{Path to a depth output file (default NULL).}

\item{chromosomes_names_file_path}{Path to a contig names file (default NULL).}

\item{output.file}{Path to an output file in PNG format. If NULL, the plot will be drawn in the default graphic device (default: NULL).}

\item{width}{Width of the output file if specified, in inches (default: 12).}

\item{height}{Height of the output file if specified, in inches (default: 4 * number of plots).}

\item{dpi}{Resolution of the output file if specified, in dpi (default: 300).}

\item{tracks}{Tracks to be plotted. Possible values are "position_fst", "window_fst", "window_snp_males",
"window_snp_females", "window_snp_combined", "window_snp_ratio", "depth_males", "depth_females", "depth_combined", "depth_ratio"
(default: c("window_fst", "window_snp_combined", "depth_ratio")).}

\item{point.size}{Size of the points in the plot (default: 0.1).}

\item{depth.type}{Type of depth to be plotted, either "absolute" or "relative" (default: "absolute").}

\item{min.depth}{Minimum depth to compute depth ratio.
The log of ratio for positions with depth lower than this value in either sex will be 0 (default: 10).}

\item{depth.ratio.lines}{If TRUE, lines will be drawn for ratios of 2, 3/2, 2/3, and 1/2 (default: FALSE).}

\item{major.lines.y}{If TRUE, major grid lines will be plotted for the y axis (default: TRUE).}

\item{major.lines.x}{If TRUE, major grid lines will be plotted for the y axis (default: TRUE).}

\item{fst.window.color}{Color of the FST window track (default: "grey50").}

\item{males.color}{Color of the male window tracks (default: "dodgerblue3").}

\item{females.color}{Color of the female window tracks (default: "firebrick2").}

\item{alpha.value}{Transparency values for combined tracks (default: 0.25).}
}
\description{
Generate plots for a specific scaffold and/or region from the results of PSASS.
}
\examples{

}
