% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_manhattan.R
\name{plot_track_manhattan}
\alias{plot_track_manhattan}
\title{Plot a manhattan track}
\usage{
plot_track_manhattan(
  data,
  backgrounds,
  track,
  bottom.track = TRUE,
  show.chromosome.names = TRUE,
  x.labels.angle = 90,
  x.axis.title = NULL
)
}
\arguments{
\item{data}{Input data generated with the \code{\link{create_manhattan_track_data}} function)}

\item{track}{Track object storing properties for the current track, generated with the \code{\link{manhattan_track}} function}

\item{bottom.track}{If TRUE, x-axis labels and title will be added to the plot (default: TRUE)}

\item{show.chromosome.names}{If TRUE, display chromosome names on the x axis (default: TRUE)}

\item{x.labels.angle}{If TRUE, replace chromosome names with numbers for readability (default: FALSE)}

\item{x.axis.title}{Title to display on the x axis (default: NULL, i.e. no title)}
}
\value{
A ggplot object for the plot
}
\description{
Plot a single track for a manhattan plot
}
\examples{
genomic_data <- load_genome_input("psass_window.tsv")
fst_track <- manhattan_track("Fst")
manhattan_data <- create_manhattan_track_data(genomic_data, fst_track)

fst_plot <- plot_track_manhattan(manhattan_data, fst_track, bottom.track=TRUE)

}
