% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_circos_window.R
\name{track_circos_window}
\alias{track_circos_window}
\title{Circos depth track}
\usage{
track_circos_window(
  data,
  metric,
  bg.col = "white",
  point.size = 0.01,
  top.track = FALSE,
  sector.titles.expand = 1.3,
  first_sector = NULL
)
}
\arguments{
\item{data}{depth data frame.}

\item{bg.col}{Background color for sectors, either a single color or a vector of colors for each sector (default "white").}

\item{point.size}{Size of a point in the plot (default 0.01).}

\item{top.track}{If TRUE, this track will be considered top track of the plot and the x axis will be drawn (default FALSE).}

\item{pool}{Pool to draw the track for, either 1 or 2.}

\item{type}{Type of depth to draw the track for, either "absolute" or "relative" (default "absolute").}

\item{sector.names}{Vector of contig names obtained with the \code{\link{load_contig_names}} (default NULL).}

\item{sector.title.expand}{Value controlling the distance between sector titles and the top axis (default 1.3).}

\item{sectors}{Vector with the names of the sectors in the plot (default NULL).}

\item{pools.color}{Color for each pool (default c("firebrick2", "dodgerblue3"))}
}
\description{
Draws a track on a circos plot for depth data. This function is intended for use in the \code{\link{draw_circos_plot}} function.
}
