% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_circos_window_fst.R
\name{track_circos_window_fst}
\alias{track_circos_window_fst}
\title{Circos FST window track}
\usage{
track_circos_window_fst(
  data,
  ylim = c(0, 1.025 * max(data$Fst) + 0.01),
  bg.col = "white",
  point.size = 0.01,
  top.track = FALSE,
  sector.names = NULL,
  sector.titles.expand = 1.3,
  sectors = NULL,
  points.color = "grey20",
  points.color.unplaced = c("dodgerblue3", "goldenrod1")
)
}
\arguments{
\item{data}{FST window data frame.}

\item{ylim}{Limits of the y axis (default c(0, 1.025 * max(FST))).}

\item{bg.col}{Background color for sectors, either a single color or a vector of colors for each sector (default "white").}

\item{point.size}{Size of a point in the plot (default 0.01).}

\item{top.track}{If TRUE, this track will be considered top track of the plot and the x axis will be drawn (default FALSE).}

\item{sector.names}{Vector of contig names obtained with the \code{\link{load_contig_names}} (default NULL).}

\item{sectors}{Vector with the names of the sectors in the plot (default NULL).}

\item{points.color}{Color of points for fst_pos, fst_win, and depth_ratio tracks (default: "grey20)}

\item{points.color.unplaced}{Alternating point color for scaffolds when plotting unplaced scaffolds (default: c("dodgerblue3", "goldenrod1"))}

\item{sector.title.expand}{Value controlling the distance between sector titles and the top axis (default 1.3).}
}
\description{
Draws a track on a circos plot for FST window data. This function is intended for use in the \code{\link{draw_circos_plot}} function.
}
