% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_circos_window_snp.R
\name{track_circos_window_snp}
\alias{track_circos_window_snp}
\title{Circos SNP window track}
\usage{
track_circos_window_snp(
  data,
  pool,
  ylim = 0,
  bg.col = "white",
  point.size = 0.01,
  top.track = FALSE,
  sector.names = NULL,
  sector.titles.expand = 1.3,
  sectors = NULL,
  pools.color = c("firebrick2", "dodgerblue3")
)
}
\arguments{
\item{data}{SNP window data frame.}

\item{pool}{Pool to draw the track for, either 1 or 2}

\item{ylim}{Limits of the y axis (default c(0, 1.025 * max(data$Males, data$Females))).}

\item{bg.col}{Background color for sectors, either a single color or a vector of colors for each sector (default "white").}

\item{point.size}{Size of a point in the plot (default 0.01).}

\item{top.track}{If TRUE, this track will be considered top track of the plot and the x axis will be drawn (default FALSE).}

\item{sector.names}{Vector of contig names obtained with the \code{\link{load_contig_names}} (default NULL).}

\item{sectors}{Vector with the names of the sectors in the plot (default NULL).}

\item{pools.color}{Color of points for each pool (default: c("firebrick2", "dodgerblue3") for pool1 and pool2 respectively)}

\item{sector.title.expand}{Value controlling the distance between sector titles and the top axis (default 1.3).}
}
\description{
Draws a track on a circos plot for SNP window data. This function is intended for use in the \code{\link{draw_circos_plot}} function.
}
