% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_scaffold_window_depth_ratio.R
\name{track_scaffold_window_depth_ratio}
\alias{track_scaffold_window_depth_ratio}
\title{Scaffold ratio depth window track}
\usage{
track_scaffold_window_depth_ratio(data, scaffold.name, region,
  type = "absolute", min.depth = 0, major.lines.y = TRUE,
  major.lines.x = FALSE, ylim = NULL, males.color = "dodgerblue3",
  females.color = "firebrick2", bottom.track = FALSE,
  ratio.lines = FALSE)
}
\arguments{
\item{data}{depth window data frame.}

\item{scaffold.name}{Name of the plotted scaffold (for the x axis)}

\item{region}{A vector specifying the boundaries of the region to be plotted, e.g. c(125000, 250000).}

\item{type}{Type of depth to use, either "absolute" or "relative" (default: "absolute").}

\item{min.depth}{Minimum depth to compute depth ratio (default: 0).}

\item{major.lines.y}{If TRUE, major grid lines will be plotted for the y axis (default: TRUE).}

\item{major.lines.x}{If TRUE, major grid lines will be plotted for the y axis (default: TRUE).}

\item{ylim}{Limits of the y axis (default: c(min(FST), 1)).}

\item{males.color}{Color of the plotted area for males (default: "dodgerblue3").}

\item{females.color}{Color of the plotted area for females (default: "firebrick2").}

\item{bottom.track}{If TRUE, this track will be considered bottom track of the plot and the x axis will be drawn (default: FALSE).}

\item{ratio.lines}{If TRUE, lines will be drawn for ratios of 2, 3/2, 2/3, and 1/2 (default: FALSE).}
}
\description{
Draws a scaffold plot track for the ratio of depth window data.
This function is intended for use in the \code{\link{draw_scaffold_plot}} function.
}
