% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_psass_output.R
\name{detect_chromosomes}
\alias{detect_chromosomes}
\title{Detect chromosomes}
\usage{
detect_chromosomes(data)
}
\arguments{
\item{data}{Psass output data frame}
}
\value{
A named vector with contig IDs as names and chromosome names as values
}
\description{
Automatically detect chromosomes in a psass output data frame
}
\examples{
chromosomes <- detect_chromosomes(data)
}
