% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_circos_plot.R
\name{draw_circos_plot}
\alias{draw_circos_plot}
\title{Draw circos plot}
\usage{
draw_circos_plot(
  data,
  output.file = NULL,
  width = 2400,
  height = 2400,
  res = 120,
  tracks = c("fst_win", "snp_win_pool1", "snp_win_pool2", "depth_ratio"),
  highlight = NULL,
  zoom.highlights = FALSE,
  zoom.ratio = 2,
  zoom.suffix = " (zoom)",
  base.color = "white",
  highlight.color = "grey80",
  point.size = 0.1,
  points.color = "grey20",
  pools.color = c("firebrick2", "dodgerblue3"),
  color.unplaced = FALSE,
  points.color.unplaced = c("dodgerblue3", "goldenrod1"),
  sector.titles.expand = NULL,
  depth.type = "absolute",
  min.depth = 10
)
}
\arguments{
\item{data}{A PSASS data structure obtained with the \code{\link{load_data_files}} function.}

\item{output.file}{Path to an output file in PNG format. If NULL, the plot will be drawn in the default graphic device (default: NULL).}

\item{width}{Width of the output file if specified, in pixels (default: 2400).}

\item{height}{Height of the output file if specified, in pixels (default: 2400).}

\item{res}{Resolution of the output file if specified, in \% (default: 120).}

\item{tracks}{Tracks to be plotted. Possible values are "fst_pos", "fst_win", "snp_pos", "snp_win_pool1",
"snp_win_pool2", "snp_win", "depth_pool1", "depth_pool2", "depth_ratio"
(default: c("fst_win", "snp_win_pool1", "snp_win_pool2", "depth_ratio")).}

\item{highlight}{A vector of sectors to highlight, for instance c("LG5") or c("NC_02536.1", "NC_02543.1") (default: NULL).}

\item{zoom.highlights}{If TRUE, highlighted sectors will be enlarged and placed at the top of the plot (default: FALSE).}

\item{zoom.ratio}{Zoom factor for highlighted sectors if zoom.highlights is TRUE (default: 2).}

\item{zoom.suffix}{Suffix to append to the name of zoomed highlighted sectors (default: " (zoom)").}

\item{base.color}{Background color of a non-highlighted sector (default: "white").}

\item{highlight.color}{Background color of a highlighted sector (default: "grey80").}

\item{point.size}{Size of the points in the plot (default: 0.1).}

\item{points.color}{Color of points for fst_pos, fst_win, and depth_ratio tracks (default: "grey20)}

\item{pools.color}{Color of points for each pool for snp_pos, snp_win_pool1, snp_win_pool2, depth_pool1, and depth_pool2 tracks
(default: c("firebrick2", "dodgerblue3") for pool1 and pool2 respectively)}

\item{color.unplaced}{If TRUE, unplaced scaffolds will be colored with alternating colors, like in a manhattan plot (default: FALSE)}

\item{points.color.unplaced}{Alternating point color for scaffolds when plotting unplaced scaffolds (default: c(dodgerblue3", "goldenrod1"))}

\item{sector.titles.expand}{Parameter to manually override the space between sector titles and x-axis (default: NULL).}

\item{depth.type}{Type of depth to be plotted, either "absolute" or "relative" (default: "absolute").}

\item{min.depth}{Minimum depth to compute depth ratio.
The ratio for positions with depth lower than this value in either sex will be 1 (default: 10).}
}
\description{
Draw a circos plot from the PSASS data
}
\examples{

draw_circos_plot(data, tracks = c("position_fst", "position_snp", "depth_males"),
                 highlight = c("Chr10", "Chr17"), color.unplaced = TRUE)
}
