% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_manhattan.R
\name{draw_manhattan_plot}
\alias{draw_manhattan_plot}
\title{Draw manhattan plot}
\usage{
draw_manhattan_plot(
  data,
  contig.lengths,
  tracks,
  output.file = NULL,
  width = 14,
  track.height = 6,
  res = 300,
  chromosomes.as.numbers = FALSE,
  show.chromosome.names = TRUE,
  x.axis.title = NULL,
  default.point.color = c("dodgerblue3", "darkgoldenrod2"),
  default.bg.color = c("grey85", "white"),
  default.point.size = 0.5,
  default.ylim = NULL
)
}
\arguments{
\item{data}{Genomic data (e.g. result of PSASS or RADSex loaded with the \code{\link{load_genome_input}} function)}

\item{contig.lengths}{Contig lengths vector (e.g. from the output of the \code{\link{load_genome_input}} function)}

\item{tracks}{List of tracks to plot. Tracks can be generated with the \code{\link{manhattan_track}} function}

\item{output.file}{Path to an output file for the generated manhattan plot, or NULL to plot in the current R device (default: NULL)}

\item{width}{Plot width when plotting to an output file, in inches (default: 12)}

\item{track.height}{Height of a single track when plotting to an output file, in inches (default: 6)}

\item{res}{Image resolution when plotting to an output file, in dpi (default: 300)}

\item{chromosomes.as.numbers}{If TRUE, replace chromosome names with numbers for readability (default: FALSE)}

\item{show.chromosome.names}{If TRUE, display chromosome names on the x axis (default: TRUE)}

\item{x.axis.title}{Title to display on the x axis (default: NULL, i.e. no title)}

\item{default.point.color}{Default color for a track when not specified in track data,
either a string (e.g. "grey20") or a vector of alternating colors (e.g. c("red", "blue") for two colors) (default: c("dodgerblue3", "darkgoldenrod2"))}

\item{default.bg.color}{Default background color when not specified in track data,
either a string (e.g. "grey20") or a vector of alternating colors (e.g. c("grey80", "white") for two colors) (default: c("grey85", "white"))}

\item{default.point.size}{Default point size for a track when not specified in track data (default: 0.5)}

\item{default.ylim}{Default y-axis limits for a track when not specified in track data (default: NULL, i.e. infer from data)}
}
\value{
Combined plot data (ggplot object)
}
\description{
Generate a manhattan plot with multiple tracks for the entire genome
}
\examples{
# Manhattan plot showing female-specific and male-specific SNPs on two tracks
genomic_data <- load_genome_input("psass_window.tsv")
draw_manhattan_plot(genomic_data$data, genomic_data$lengths,
                    tracks = list(manhattan_track("Snps_females", point.color = c("firebrick1", "firebrick3")),
                                  manhattan_track("Snps_males", point.color = c("dodgerblue1", "dodgerblue3"))),
                    output.file = "manhattan.png",
                    chromosomes.as.numbers = TRUE, show.chromosome.names = TRUE, x.axis.title = "Chromosome")

}
