% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_scaffold_plot.R
\name{draw_scaffold_plot}
\alias{draw_scaffold_plot}
\title{Draw scaffold plot}
\usage{
draw_scaffold_plot(data, scaffold, region = NULL, output.file = NULL,
  width = 12, height = 4, dpi = 300, tracks = c("window_fst",
  "window_snp_males", "window_snp_females", "coverage_ratio"),
  point.size = 0.5, coverage.type = "absolute", min.coverage = 10,
  major.lines.y = TRUE, major.lines.x = FALSE,
  fst.window.color = "grey50", males.color = "dodgerblue3",
  females.color = "firebrick2")
}
\arguments{
\item{data}{A PoolSex data structure obtained with the \code{\link{load_data_files}} function.}

\item{scaffold}{Name of the scaffold to be plotted. The name can be from the dataset (e.g. "NC_0245486.2") or from the
'chromosomes_names' file (e.g. "LG4").}

\item{region}{A vector specifying the boundaries of the region to be plotted, e.g. c(125000, 250000). If NULL, the entire scaffold
will be plotted (default: NULL).}

\item{output.file}{Path to an output file in PNG format. If NULL, the plot will be drawn in the default graphic device (default: NULL).}

\item{width}{Width of the output file if specified, in inches (default: 12).}

\item{height}{Height of the output file if specified, in inches (default: 4 * number of plots).}

\item{dpi}{Resolution of the output file if specified, in dpi (default: 300).}

\item{tracks}{Tracks to be plotted. Possible values are "position_fst", "window_fst", "position_snp", "window_snp_males",
"window_snp_females", "combined_snp", "coverage_males", "coverage_females", "coverage_ratio"
(default: c("window_fst", "combined_snp", "coverage_ratio")).}

\item{point.size}{Size of the points in the plot (default: 0.1).}

\item{coverage.type}{Type of coverage to be plotted, either "absolute" or "relative" (default: "absolute").}

\item{min.coverage}{Minimum coverage to compute coverage ratio.
The log of ratio for positions with coverage lower than this value in either sex will be 0 (default: 10).}
}
\description{
Draw a scaffold plot from the PoolSex data
}
\examples{

}
