% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos_genome.R
\name{plot_genome_circos}
\alias{plot_genome_circos}
\title{Plot genome circos}
\usage{
plot_genome_circos(
  contig_lengths_file_path,
  plot.unplaced = TRUE,
  prefix = NULL,
  window_fst_file_path = NULL,
  position_fst_file_path = NULL,
  window_snps_file_path = NULL,
  position_snps_file_path = NULL,
  depth_file_path = NULL,
  chromosomes_names_file_path = NULL,
  output.file = NULL,
  width = 2400,
  height = 2400,
  res = 120,
  tracks = c("fst_win", "snp_win_pool1", "snp_win_pool2", "depth_ratio"),
  highlight = NULL,
  zoom.highlights = FALSE,
  zoom.ratio = 2,
  zoom.suffix = " (zoom)",
  base.color = "white",
  highlight.color = "grey80",
  point.size = 0.1,
  points.color = "grey20",
  pools.color = c("firebrick2", "dodgerblue3"),
  color.unplaced = FALSE,
  points.color.unplaced = c(`0` = "dodgerblue3", `1` = "goldenrod1"),
  sector.titles.expand = NULL,
  depth.type = "absolute",
  min.depth = 10
)
}
\arguments{
\item{contig_lengths_file_path}{Path to a contig lengths file.}

\item{plot.unplaced}{If TRUE, unplaced contigs will be plotted as a supercontig (default TRUE).}

\item{prefix}{Prefix (including full path) to a complete dataset. If prefix is specified, it will be override individual file specifications
such as "window_fst_file_path" (default NULL).}

\item{window_fst_file_path}{Path to a FST window output file (default NULL).}

\item{position_fst_file_path}{Path to a FST positions output file (default NULL).}

\item{window_snps_file_path}{Path to a SNPs window output file (default NULL).}

\item{position_snps_file_path}{Path to a SNPs positions output file (default NULL).}

\item{depth_file_path}{Path to a depth output file (default NULL).}

\item{chromosomes_names_file_path}{Path to a contig names file (default NULL).}

\item{output.file}{Path to an output file in PNG format. If NULL, the plot will be drawn in the default graphic device (default: NULL).}

\item{width}{Width of the output file if specified, in pixels (default: 2400).}

\item{height}{Height of the output file if specified, in pixels (default: 2400).}

\item{res}{Resolution of the output file if specified, in \% (default: 120).}

\item{tracks}{Tracks to be plotted. Possible values are "fst_pos", "fst_win", "snp_pos", "snp_win_pool1",
"snp_win_pool2", "snp_win", "depth_pool1", "depth_pool2", "depth_ratio"
(default: c("fst_win", "snp_win_pool1", "snp_win_pool2", "depth_ratio")).}

\item{highlight}{A vector of sectors to highlight, for instance c("LG5") or c("NC_02536.1", "NC_02543.1") (default: NULL).}

\item{zoom.highlights}{If TRUE, highlighted sectors will be enlarged and placed at the top of the plot (default: FALSE).}

\item{zoom.ratio}{Zoom factor for highlighted sectors if zoom.highlights is TRUE (default: 2).}

\item{zoom.suffix}{Suffix to append to the name of zoomed highlighted sectors (default: " (zoom)").}

\item{base.color}{Background color of a non-highlighted sector (default: "white").}

\item{highlight.color}{Background color of a highlighted sector (default: "grey80").}

\item{point.size}{Size of the points in the plot (default: 0.1).}

\item{points.color}{Color of points for fst_pos, fst_win, and depth_ratio tracks (default: "grey20)}

\item{pools.color}{Color of points for each pool for snp_pos, snp_win_pool1, snp_win_pool2, depth_pool1, and depth_pool2 tracks
(default: c("dodgerblue3", "firebrick2") for pool1 and pool2 respectively)}

\item{color.unplaced}{If TRUE, unplaced scaffolds will be colored with alternating colors, like in a manhattan plot (default: FALSE)}

\item{points.color.unplaced}{Alternating point color for scaffolds when plotting unplaced scaffolds (default: c("0"="dodgerblue3", "1"="goldenrod1"))}

\item{sector.titles.expand}{Parameter to manually override the space between sector titles and x-axis (default: NULL).}

\item{depth.type}{Type of depth to be plotted, either "absolute" or "relative" (default: "absolute").}

\item{min.depth}{Minimum depth to compute depth ratio.
The ratio for positions with depth lower than this value in either sex will be 1 (default: 10).}
}
\description{
Generate a circos plot of the entire genome from the results of PSASS.
}
\examples{

# Standard plot with default options
plot_genome_circos(contig_lengths_file_path = 'data/contig_lengths.tsv',
                   chromosomes_names_file_path = 'data/chromosomes_names.tsv',
                   prefix = 'data/psass',
                   output.file = 'figures/psass_genome.png')

# Plot FST positions and SNP window for the first pool with highlight on NC_02456.3
plot_genome_circos(contig_lengths_file_path = 'data/contig_lengths.tsv', prefix = 'data/psass',
                   tracks = c("fst_pos", "snp_win_pool1"), highlight = c("NC_02456.3"))
}
